#8 [RENAME] rename regex, image launcher file
