+++ core/src/main/java/com/sopt/core/extension/ContextExt.kt
@@ -1,7 +1,11 @@
 package com.sopt.core.extension
 
 import android.content.Context
+import android.os.Build
 import android.widget.Toast
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.PickVisualMediaRequest
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.StringRes
 
 fun Context.toast(@StringRes message: Int) {
@@ -18,4 +22,17 @@ fun Context.longToast(@StringRes message: Int) {
 
 fun Context.stringOf(@StringRes message: Int): String {
     return getString(message)
+}
+
+fun Context.launchImagePicker(
+    galleryLauncher: ActivityResultLauncher<String>,
+    photoPickerLauncher: ActivityResultLauncher<PickVisualMediaRequest>
+) {
+    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
+        // API 33 미만: 갤러리 실행
+        galleryLauncher.launch("image/*")
+    } else {
+        // API 33 이상: 포토피커 실행 (이미지만 선택)
+        photoPickerLauncher.launch(PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly))
+    }
 }
