+++ presentation/src/main/java/com/sopt/presentation/groupCreate/groupCreateSuccess/regex/Regex.kt
@@ -1,0 +1,8 @@
+package com.sopt.presentation.groupCreate.groupCreateSuccess.regex
+
+class Regex {
+    fun generateRandomCode(): String {
+        val charset = ('0'..'9') + ('A'..'Z')
+        return List(6) { charset.random() }.joinToString("")
+    }
+}
