+++ presentation/src/main/java/com/sopt/presentation/groupCreate/permission/ImagePickerLaunchers.kt
@@ -1,0 +1,25 @@
+package com.sopt.presentation.groupCreate.permission
+
+import android.net.Uri
+import androidx.activity.compose.ManagedActivityResultLauncher
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.PickVisualMediaRequest
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.runtime.Composable
+
+class ImagePickerLaunchers {
+    @Composable
+    fun rememberGalleryLauncher(onImageSelected: (Uri) -> Unit): ActivityResultLauncher<String> {
+        return rememberLauncherForActivityResult(ActivityResultContracts.GetContent()) { uri: Uri? ->
+            uri?.let(onImageSelected)
+        }
+    }
+
+    @Composable
+    fun rememberPhotoPickerLauncher(onImageSelected: (Uri) -> Unit): ManagedActivityResultLauncher<PickVisualMediaRequest, Uri?> {
+        return rememberLauncherForActivityResult(ActivityResultContracts.PickVisualMedia()) { uri: Uri? ->
+            uri?.let(onImageSelected)
+        }
+    }
+}
